+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
 - System wide default presentation ([#119], [#120])
+- Co-owner room role, permissions to view and edit all rooms (incl. memberships, files and settings) ([#110], [#116])
 
 ## [1.1.0] - 2021-03-08
 ### Added
@@ -108,6 +110,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#120]: https://github.com/THM-Health/PILOS/pull/120
 [#121]: https://github.com/THM-Health/PILOS/issues/121
 [#122]: https://github.com/THM-Health/PILOS/pull/122
+[#110]: https://github.com/THM-Health/PILOS/issues/110
+[#116]: https://github.com/THM-Health/PILOS/pull/116
 
 [unreleased]: https://github.com/THM-Health/PILOS/compare/v1.1.0...HEAD
 [1.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v1.0.0